Align weight dtype in DISABLED grad_input path under autocast
Float8Linear.forward casts only the input to the autocast dtype, so
save_for_backward stores a bf16 input next to an fp32 weight. The
forward gemm survives because autocast is still active there and
torch.mm is autocast-eligible. Backward runs with autocast off, so when
cast_config_weight_for_grad_input is ScalingType.DISABLED the raw fp32
weight reaches torch.mm against a bf16 grad_output and it raises
"expected mat1 and mat2 to have the same dtype".

Cast the saved weight to the dtype of the grad_output tensor it is
multiplied against. This is a no-op without autocast, where both are
already fp32. The grad_weight gemm needs no change, input_hp was
already cast to the autocast dtype in forward.

Adds test_autocast_backward_scaling_disabled, which uses an all-DISABLED
config so it runs without sm_89 and checks both gradients against a
plain nn.Linear reference under the same autocast.

Fixes #4616

+++ torchao/float8/float8_linear.py
@@ -126,7 +126,12 @@ class matmul_with_hp_or_float8_args(torch.autograd.Function):
             # TODO(future PR): var name is axiswise specific, fix it
             weight_t_maybe_fp8_dim0 = weight_hp_t
         elif c.cast_config_weight_for_grad_input.scaling_type is ScalingType.DISABLED:
-            weight_t_maybe_fp8_dim0 = weight_hp_t
+            # under autocast, `grad_output` is in the autocast dtype but the saved
+            # weight is still in the original high precision dtype, and `torch.mm`
+            # in backward does not autocast. Align the dtypes here.
+            weight_t_maybe_fp8_dim0 = weight_hp_t.to(
+                grad_output_reshaped_maybe_fp8_dim0.dtype
+            )
         else:
             weight_t_maybe_fp8_dim0 = hp_tensor_to_float8_dynamic(
                 weight_hp_t,
